src/requirements.txt
attrs==23.2.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
filelock==3.13.1
flasgger==0.9.7.1
flask-redis==0.4.0
flask-swagger-ui==4.11.1
fsspec==2023.12.2
greenlet==3.0.3
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mistune==3.0.2
mpmath==1.3.0
mysql-connector-python==8.0.28
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.0
pycparser==2.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
schedule==1.2.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==1.4.31
sympy==1.12
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
Werkzeug==2.2.2
wheel==0.41.3
slackclient==2.9.4
# jwt==1.3.1
Flask==2.0.0
Flask-JWT-Extended==4.3.1
PyJWT==2.1.0